DataCleanup Job dynamic retention days and deletion
Problems Fixed:
1) If a store with data retention is deleted, it incorrectly holds
the lock there by prevening other jobs from running.
2) If a store's data retention is modified, it requires the cluster
to be bounced as the retention day is read at the start and never
altered.

Problem not fixed:
 If the data cleanup job retention frequency is set for the store
and if the frequency is modified, it would still require a cluster
bounce.

Fix:
   1) Instead of taking the store retention time, data cleanup job
takes in the store name and metadata store and computes the retention
time dynamically for each run.
   2) When a store can't be retrieved from the metadata store the
cleanup job is skipped.
   3) All the code after the lock acquistion is moved inside the
try/finally block. Previously beginBatchModification was outside
of the lock, which threw exception and caused the lock to be not
freed.
   4) Added unit tests for both the problems fixed.