Add idle connection timeout for Client connections
If Voldemort client lives behind a firewall, the connections
could be dropped by the firewall silently. The firewall also drops any
future packets sent on the connection, which causes lot of timeout for
low throughput voldemort clients.

The fix adds the timeout to the client config, by default the idle
connection timeout is disabled.