When store is missing, error message is not clear
When a store is missing on the Server, Voldemort error message on
the client used to say

Failed to read metadata key:"XXX" delete config/.temp config/.version directories and restart.

Now it says store XXX does not exist on node YY

This will be easier to reason from the client perspective.