Add method to make re-use of the AdminClient easier
When AdminClient is created newly for each operation, AdminClient needs to
re-establish every connection to the Voldemort Server. This makes AdminClient
operations take a longer period of time.

But if AdminClient is reused across cluster modification, then AdminClient
will cause inconsistent operations. This new method will help the caller
to identify if the cached AdminClient is still valid and can be re-used.