AdminClientPool, to easily pool AdminClient
AdminClientPool is added for managing the pools of AdminClient.

AdminClient unlike StoreClient can't be used across Cluster
modifications. So previously AdminClient needs to be created
every time. this was costly as the connections need to be re-established
every time.

AdminClientPool solves this problem by discarding AdminClient if cluster
is modified.

AdminClientPool still does not solve the problem of failing operation
during cluster modification. But it will work correctly after the
cluster is modified.