Fetch Single Store only for BnP Store creation
BnP supports option for querying only the current store.
By default it queries full stores.xml as this option requires
some server side changes that will go in 1.10.18

But once the server side changes are deployed, using this on
a cluster with large number of stores will speed up the
pre-processing.