Validation and Generate Script Utility added for testing purpose
The GenereateScript is not secure  and it depends on the user
to give sanitized input and reasonable script. For the same reason
it just generates the output file, which must be manually reviewed
before executing the script. GenerateScript is more of use it at
your own risk.

ValidateNodeIdCLI helps to validate the auto detect node Id, before
removing the node Id from all the configs.

ValidateNodeIdCLI when combined witht he GenerateScript, can be used
to validate the result of entire cluster, safely before removing
all the node ids.