Rest Server Port is not serialized correctly
Problems
1) While running ./gradlew clean build, exits the process and the
build fails in the middle.
2) When Voldemort server rest validation fails it exits the process.
3) Cluster does not serialize the rest port correctly, which caused the
rest port validation to fail.
4) Before auto node detection the tests were using in memory
cluster instead of the cluster in the metadata. Now both tests and
product code use the same code path, which caused the tests to fail.

Fix:
1) Cluster serializes the rest port, if it is greater than zero.
2) When rest server validation fails, it throws an exception, instead
of exiting the process. (Searched code for System.exit and Coordinator
Server does the same, but saving that for a different day).
3) Node state string contains the rest port, if it is present.
4) Let the RestServiceR2StoreTest fail with an actual error message,
instead of boiler plate error message, which made the debugging harder.