RO Store Create floods the Log with error messages
Creating a RO store queries for an existing store, which
fails with StoreNotFoundException. This exception is logged with call
stack, which floods the logs on every store creation.

This may trick the alerting system into treating this as error.
Not logging a call stack and reducing the log to info, when such
exceptions are logged.