Data Cleanup job Does not run on system stores
1) Client registry System store is a in-memory store and supposed to be cleaned up after 7 days.
Last change to the DataCleanupJob made the system stores fail with the missing store exception.

Clients re-use the same client id, so unless lots of clients become
dead and removed, this will not cause a leak on the server resources. The effect is negligible.

Now the DataCleanupJob checks for both system stores and normal stores for a store definition.

2) If the store retention days is modified to zero, then the store will
delete all the records. But if the store is started with 0 retention days
it means the data retention is not enabled. Fixed the discrepancy.